Reformat settings menu for admin and superuser
