+++ app/femr/ui/views/admin/configure/manage.scala.html
@@ -8,7 +8,7 @@
     }
 
 @admin("Configure", currentUser, styles = additionalStyles) {
-    @helper.form(action = ConfigureController.indexPost()) {
+    @helper.form(action = ConfigureController.managePost()) {
         <table id="configurationTable">
             <tr>
                 <th>Feature</th>
