+++ app/femr/ui/views/admin/inventory/manage.scala.html
@@ -15,8 +15,8 @@
         <script type = "text/javascript" src="@routes.Assets.versioned("js/admin/inventory.js")"></script>
     }
 
-@admin("Inventory Tracking", currentUser, styles = additionalStyles, scripts = additionalScripts) {
-    @helper.form(action = InventoryController.indexPost()) {
+@admin("Inventory", currentUser, styles = additionalStyles, scripts = additionalScripts) {
+    @helper.form(action = InventoryController.managePost()) {
 
         <div class="newMedicationWrap">
