+++ app/femr/ui/views/partials/authenticated.scala.html
@@ -6,7 +6,7 @@
 @import femr.ui.controllers.routes.TriageController
 @import femr.ui.controllers.routes.MedicalController
 @import femr.ui.controllers.routes.PharmaciesController
-@import femr.ui.controllers.routes.SuperuserController
+@import femr.ui.controllers.superuser.routes.SuperuserController
 @import femr.ui.controllers.admin.routes.AdminController
 @import femr.ui.controllers.routes.ResearchController
 <div class="navigationLogo">
