+++ conf/routes
@@ -2,13 +2,14 @@
 # This file defines all application routes (Higher priority routes first)
 # ~~~~
 GET         /*path/                                            femr.ui.controllers.ApplicationController.removeTrailingSlash(path: String)
-GET         /superuser/tabs/:name                              @femr.ui.controllers.SuperuserController.contentGet(name: String)
-POST        /superuser/tabs/:name                              @femr.ui.controllers.SuperuserController.contentPost(name: String)
-GET         /superuser/tabs                                    @femr.ui.controllers.SuperuserController.tabsGet()
-POST        /superuser/tabs                                    @femr.ui.controllers.SuperuserController.tabsPost()
-GET         /superuser/trips                                   @femr.ui.controllers.SuperuserController.tripsGet()
-POST        /superuser/trips                                   @femr.ui.controllers.SuperuserController.tripsPost()
-GET         /superuser                                         @femr.ui.controllers.SuperuserController.indexGet()
+GET         /superuser/tabs/:name                              @femr.ui.controllers.superuser.TabController.fieldsGet(name: String)
+POST        /superuser/tabs/:name                              @femr.ui.controllers.superuser.TabController.fieldsPost(name: String)
+GET         /superuser/tabs                                    @femr.ui.controllers.superuser.TabController.manageGet()
+POST        /superuser/tabs                                    @femr.ui.controllers.superuser.TabController.managePost()
+GET         /superuser/trips                                   @femr.ui.controllers.superuser.TripController.manageGet()
+GET         /superuser/trips/menu                              @femr.ui.controllers.superuser.TripController.menuGet()
+POST        /superuser/trips/menu                              @femr.ui.controllers.superuser.TripController.menuPost()
+GET         /superuser                                         @femr.ui.controllers.superuser.SuperuserController.indexGet()
 #Admin
 POST        /admin/users/create                                @femr.ui.controllers.admin.UsersController.createPost()
 GET         /admin/users/create                                @femr.ui.controllers.admin.UsersController.createGet()
@@ -16,12 +17,12 @@ POST        /admin/users/edit                                  @femr.ui.controll
 GET         /admin/users/edit/:id                              @femr.ui.controllers.admin.UsersController.editGet(id: Integer)
 POST        /admin/users/toggle/:id                            @femr.ui.controllers.admin.UsersController.toggleUser(id: Integer)
 GET         /admin/users                                       @femr.ui.controllers.admin.UsersController.manageGet()
-GET         /admin/inventory                                   @femr.ui.controllers.admin.InventoryController.indexGet()
-POST        /admin/inventory                                   @femr.ui.controllers.admin.InventoryController.indexPost()
+GET         /admin/inventory                                   @femr.ui.controllers.admin.InventoryController.manageGet()
+POST        /admin/inventory                                   @femr.ui.controllers.admin.InventoryController.managePost()
 POST        /admin/inventory/get                               @femr.ui.controllers.admin.InventoryController.ajaxGet()
 POST        /admin/inventory/delete/:id                        @femr.ui.controllers.admin.InventoryController.ajaxDelete(id: Integer)
-GET         /admin/configure                                   @femr.ui.controllers.admin.ConfigureController.indexGet()
-POST        /admin/configure                                   @femr.ui.controllers.admin.ConfigureController.indexPost()
+GET         /admin/configure                                   @femr.ui.controllers.admin.ConfigureController.manageGet()
+POST        /admin/configure                                   @femr.ui.controllers.admin.ConfigureController.managePost()
 GET         /admin                                             @femr.ui.controllers.admin.AdminController.index()
 #Pharmacy
 GET         /pharmacy/edit/:id                                 @femr.ui.controllers.PharmaciesController.editGet(id: Integer)
@@ -72,4 +73,4 @@ GET         /login                                             @femr.ui.controll
 GET         /                                                  @femr.ui.controllers.HomeController.index()
 
 # Map static resources from the /public folder to the /assets URL path
-GET         /assets/*file                                     controllers.Assets.versioned(path="/public", file: Asset)
+GET         /assets/*file                                      controllers.Assets.versioned(path="/public", file: Asset)
